change the way we monitor requests
Until these changes requests were counted with acceptors and monitored
like this. Now when a request is accepted we unlink it from the listener
and start to monitor it instead. Monitoring a request allows us to count
the number of active requests and eventuallly later do something with
it.

close #1