add the possibility to spawn a new request on acccept.
While the default (and generally faster) way to handle the request is to
process it  in the acceptor process, sometime you may want to spawn a
new process.

This change add the possibilty to do it by checking if the request
handler has a start_link/4 function and in that case use it to spawn a
new process and pass it the control of the socket.